Favorable
see also #2724

+++ database/migrations/2019_10_05_141513_alter_table_contentsets_add_redirect_url.php
@@ -26,7 +26,9 @@ class AlterTableContentsetsAddRedirectUrl extends Migration
     public function down()
     {
         Schema::table('contentsets', function (Blueprint $table) {
-            $table->dropColumn('redirectUrl');
+            if (Schema::hasColumn('contentsets', 'redirectUrl')) {
+                $table->dropColumn('redirectUrl');
+            }
         });
     }
 }
